requirements.txt
asgiref==3.8.1
Django==4.0.1
django-dotenv==1.4.2
djangorestframework==3.13.1
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.9
pycodestyle==2.12.0
PyJWT==2.8.0
python-dotenv==1.0.1
pytz==2024.1
sqlparse==0.5.1
typing_extensions==4.12.2
